% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/bracketXtract.R
\name{bracketXtract}
\alias{bracketXtract}
\title{Extract Text Between Any Brackets}
\usage{
bracketXtract(txt, br = c("(", "[", "{", "<", "all"), with = FALSE)
}
\arguments{
\item{txt}{The text string}

\item{br}{The type of bracket. Accepts \(, \[, \{, or "all"}

\item{with}{Normally set to F, if T then brackets will be included with the
  return text.}
}
\value{
This function returns all text found within brackets.
}
\description{
This is a function shamelessly pirated from the internet. It accepts a txt
string, and br, which says the types of brackets to search within.
}

