% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/buildDictionary.R
\name{buildDictionary}
\alias{buildDictionary}
\title{Build Dictionary From Text Or File}
\usage{
buildDictionary(inText, readfile = TRUE, formatKey = TRUE)
}
\arguments{
\item{inText}{The input text, or, a filename.}

\item{readfile}{If TRUE, inText is a filename to read from.}

\item{formatKey}{If TRUE, the key is formatted to <key> instead of key.}
}
\value{
This function returns a data.frame with keys and values.
}
\description{
This function extracts keys/values from file or text and return key/values in data.frame
}

