% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/buildJson.R
\name{buildJson}
\alias{buildJson}
\title{Build a JSON File For Cover Images}
\usage{
buildJson(df, returnme = FALSE, writeme = FALSE, sdir = ".",
  imagedir = "pngs/")
}
\arguments{
\item{df}{The data.frame containing authors and titles.}

\item{returnme}{If true, returns the JSON string.}

\item{writeme}{If true, writes the JSON string to a file called "covers.json2" in a directory of your choice.}

\item{sdir}{The directory to put the covers file in if write=T}

\item{imagedir}{Set to "pngs/", the place where you want the transformed files.}
}
\description{
This is a function to build a JSON file if you're planning to use the generic
image creator built by NYPL labs, called "tenprintcover". I wrote it, but I
don't currently use it any longer, so I don't plan to support it if the file
needed changes.
}
\note{
There are no tests for this function. It is not used.
}

