% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/combineFiles.R
\name{combineFiles}
\alias{combineFiles}
\title{Combine All Files In A List}
\usage{
combineFiles(listfiles, outfilename, exec = TRUE)
}
\arguments{
\item{listfiles}{The list of files to be read in. You can use "list.files()"
to generate these.}

\item{outfilename}{The name of the file, including extension, that we want to
write.}

\item{exec}{If true, the code will execute. If false, the code will print to
  a list.}
}
\value{
This function, if exec=FALSE, will return the expected output of the
  written outfile.
}
\description{
This function combines all files in a given list and outputs them into one
file. It reads in the order that listfiles is received.
}

