% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/echoDictionaryEntry.R
\name{echoDictionaryEntry}
\alias{echoDictionaryEntry}
\title{Print All Dictionary Entries Associated With a Particular Value}
\usage{
echoDictionaryEntry(mydic, key = "", val = "")
}
\arguments{
\item{mydic}{A data.frame with columns "key" and "val"}

\item{key}{A search value for the "key" column}

\item{val}{A search value for the "val" column}
}
\value{
This function returns a vector of response values
}
\description{
This function returns the values or keys associated with the opposite value entered, e.g., if you enter a key, the function will return all of its values, and vice versa.
}
\examples{
testdic <- data.frame(key=c("<key1>", "<key2>"), val=c("a", "b"), stringsAsFactors=F)
echoDictionaryEntry(testdic, "key1")
}

