% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/extractKey.R
\name{extractKey}
\alias{extractKey}
\title{Extract A Key From The Dictionary}
\usage{
extractKey(inString, format = TRUE)
}
\arguments{
\item{inString}{A string to be formatted.}

\item{format}{If set to TRUE (default), then <> are added around the word for
insertion into a story.}
}
\value{
This function returns the modified string.
}
\description{
A function to extract keys from the unique dictionary syntax, which goes key
::= val1 | val2 | etc..., and put it in its proper format.
}

