% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/extractValue.R
\name{extractValue}
\alias{extractValue}
\title{Extract A Set Of Values From The Dictionary}
\usage{
extractValue(inString, format = TRUE)
}
\arguments{
\item{inString}{A string to be formatted.}

\item{format}{If set to TRUE (default), then the string is split into individual members by | before being returned.}
}
\value{
This function returns the modified string.
}
\description{
A function to extract values from the unique dictionary syntax, which goes
key ::= val1 | val2 | etc..., and put it in its proper format, which is
individual members of a vector.
}

