% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/generateCharacters.R
\name{generateCharacters}
\alias{generateCharacters}
\title{Generate Character Names By Gender}
\usage{
generateCharacters(charnames = c("cFemName", "cMaleName", "cFriendName",
  "cPetName"), chargenders = c("F", "M", "F", "M"), whichnames = "first",
  writefile = FALSE, outname = "aNames.txt")
}
\arguments{
\item{charnames}{A vector of keys to use for the characters, e.g., cFemName}

\item{chargenders}{A vector of "M" or "F", representing male or female.}

\item{whichnames}{Default is first, set to last or both to change the output
string}

\item{writefile}{If set to true, this function will write a file.}

\item{outname}{The directory and name of file that we should write if
  writefile=TRUE}
}
\value{
This function returns the value of the file that it will write.
}
\description{
This function generates character first names by gender. It requires the
randomNames package.
}

