% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/grammaR.R
\docType{package}
\name{grammaR}
\alias{grammaR}
\alias{grammaR-package}
\title{Package "grammaR": Recursively generated text and story writing with R}
\description{
This package was built to help a hobby writer write things better using code.
We often reuse phrases and words -- this package helps us generate a
dictionary and use that dictionary to vary our words. So, "he said" becomes
"he smiled" and "fear" becomes "distress", and before you know it, you have a
brand new story.
}
\details{
There are a few different functions that may be helpful to you. You can use
the \code{\link{makeNewDictionary}} command to interactively build your first
dictionary; you can use the \code{\link{makeGrammarFromStory}} function to
build a starting grammar, by paragraph, for your newest story, and you can
use \code{\link{spinStory}} to find matching values between your dictionary
and your story to spin.
}

