% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/makeCommandsCovers.R
\name{makeCommandsCovers}
\alias{makeCommandsCovers}
\title{Generate Covers With ImageMagick}
\usage{
makeCommandsCovers(df, sdir = "./", odir = "./", stockdir = "originals/",
  exec = FALSE)
}
\arguments{
\item{df}{The author/title dataframe that we'll loop through.}

\item{sdir}{The working directory, we need to be in this directory to perform
ImageMagick functions. It is much easier to set now than screw with path
later.}

\item{odir}{The out directory. As in, where you want your new files to go.}

\item{stockdir}{The directory that leads to the stock images that we'll be working with.}

\item{exec}{Set to false; set to true will attempt to execute the commands in the system() interface.}
}
\description{
This function executes ImageMagick commands to generate covers. It is a
complex function.
}
\note{
If exec=TRUE breaks/crashes R, you may need to make sure that R can find ImageMagick, and can also find ghostscript. The easiest way, on a Mac, is to use MacPorts. Install ImageMagick; if it doesn't work, install ghostscript. Consult the internet.
}

