% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/makeCoverName.R
\name{makeCoverName}
\alias{makeCoverName}
\title{Make Lowercase And Dashed Image File Name}
\usage{
makeCoverName(author, title, extension = ".jpg")
}
\arguments{
\item{author}{The first string to modify, will show up first in the output.}

\item{title}{The second string to modify, will show up second in the output.}
}
\value{
This function returns a modified string.
}
\description{
This function takes an author and a title, extracts the alphanumeric
characters, mashes them together, makes them all lowercase, hyphenates them,
and adds ".png" to the end. So, "A Midsummer Night's Dream" by "Shakespeare"
would become "shakespeare-amidsummernightsdream.png". Enjoy!
}

