% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/makeGrammarFromStory.R
\name{makeGrammarFromStory}
\alias{makeGrammarFromStory}
\title{Make a basic grammar for your story from a plain text file.}
\usage{
makeGrammarFromStory(intext)
}
\arguments{
\item{intext}{The file path to read from.}
}
\value{
This function returns a vector of character strings representing individual lines of your grammar. You will need to write them into a new file separately.
}
\description{
This function takes a text file, "intext", and returns either a full grammar separated by paragraph.
}

