% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/makeNewDictionary.R
\name{makeNewDictionary}
\alias{makeNewDictionary}
\title{Interactive: Make A New Dictionary}
\usage{
makeNewDictionary(writeme = FALSE, filename = "", carryme = c())
}
\arguments{
\item{writeme}{When active, this parameter should let you write to a file.}

\item{filename}{The file to write to, when writeme works.}

\item{carryme}{This is a vector of other lines to include in the file. It gets carried recursively through the function.}
}
\value{
This function returns a vector of key::=vals dictionary entries.
}
\description{
This is an interactive function that will help you build a dictionary. The writeme functionality is currently NOT implemented.
}

