% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/parseGrammar.R
\name{parseGrammar}
\alias{parseGrammar}
\title{Parse a grammar based on recursion}
\usage{
parseGrammar(mydic, topval = "top", currentstring = "")
}
\arguments{
\item{mydic}{The dictionary to use for parsing.}

\item{key}{The key to look up}
}
\description{
Parse a grammar based on recursion
}
\examples{
#' cat("top ::= <greeting> <place>!", "greeting ::= Hello | Howdy | Hey there | What's up", "place ::= fart <monkey>", "monkey ::= rawr | death | marquee", file="ex.data", sep="\\n")
store <- readGrammar("ex.data")
parseGrammar(store, "top")
}

