% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/readGrammar.R, R/test1.R
\name{readGrammar}
\alias{readGrammar}
\title{New Grammar Parser}
\usage{
readGrammar(filename)

readGrammar(filename)
}
\arguments{
\item{filename}{The filename to run through the function.}

\item{filename}{The filename to run through the function.}
}
\description{
New Grammar Parser

New Grammar Parser
}
\examples{
cat("top ::= <greeting> <place>!", "greeting ::= Hello | Howdy | Hey there | What's up", "place ::= fart <monkey>", "monkey ::= rawr | death | marquee", file="ex.data", sep="\\n")
store <- readGrammar("ex.data")
cat("top ::= <greeting> <place>!", "greeting ::= Hello | Howdy | Hey there | What's up", "place ::= fart <monkey>", "monkey ::= rawr | death | marquee", file="ex.data", sep="\\n")
store <- readGrammar("ex.data")
}

