% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/removeDictionaryEntry.R
\name{removeDictionaryEntry}
\alias{removeDictionaryEntry}
\title{Remove An Entry From A Dictionary Data.Frame}
\usage{
removeDictionaryEntry(mydic, key = "", val = "")
}
\arguments{
\item{mydic}{A data.frame containing a column "key" and a column "val". "key", in particular, should be formatted with < and > surrounding any key name.}

\item{key}{Initially left "", setting this will let the function search in the key column for any matching value to remove.}

\item{val}{Initially left "", setting this will let the function search in the val column for any matching value to remove.}
}
\description{
This function searches by key or val and removes any dictionary entries associated with those values. It will process the "key" before the "val", not that that should matter. Warning: This function will remove any blank entries, if you have them.
}
\examples{
mydic <- data.frame(key=c("<tigers>", "<tigers>", "<lions>", "<lions>"), val=c("a", "b", "c", "d"), stringsAsFactors=F)
removeDictionaryEntry(mydic, key="tigers")
removeDictionaryEntry(mydic, val="b")
}

