% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/spinStory.R
\name{spinStory}
\alias{spinStory}
\title{Build A Recursive Story From Normal Text}
\usage{
spinStory(infilenames, uniques, outfilenames)
}
\description{
Build A Recursive Story From Normal Text
}

