% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/writeDictionary.R
\name{writeDictionary}
\alias{writeDictionary}
\title{Write Dictionary From Data.Frame To File}
\usage{
writeDictionary(mydic, exec = FALSE, outfilename = "")
}
\arguments{
\item{mydic}{A data.frame containg "key" and "val" columns}

\item{exec}{If set to TRUE, the function will write to file at outfilename.}

\item{outfilename}{The name and path of the file you want to write, e.g., "./test.gram"}
}
\value{
This function will always return its output, regardless of whether or not exec is set to TRUE.
}
\description{
This function writes a text file based on an existing data.frame with "key" and "val" columns inside it.
}
\examples{
testdic <- data.frame(key=c("key1", "key2"), val=c("a", "b"), stringsAsFactors=F)
writeDictionary(testdic)
}

