% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cleanDocs.R
\name{cleanDocs}
\alias{cleanDocs}
\title{This Function Cleans NA From Files Before Compiling}
\usage{
cleanDocs(myPattern = ".text")
}
\arguments{
\item{myPattern}{The type of pattern to look for, e.g., .tex}
}
\value{
This function does not have an output, but overwrites files in the
  current working directory. BE CAREFUL!
}
\description{
This simple function gets all possible .tex files in the current working
directory and removes all "NA" from them, which is a side effect of some of
the other scripts.
}

