% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getUniques.R
\name{getUniques}
\alias{getUniques}
\title{Trim Data.Frame To Unique Values Only}
\usage{
getUniques(df)
}
\arguments{
\item{df}{The data.frame to check; should have a column named "val" -- that's the unique column.}
}
\value{
This function returns a data.frame with only unique values. It also should remove any stray "|" symbols.
}
\description{
This function removes all entries with a repeated value.
}

